pyproject.toml
[project]
name = "llama-index-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.0",
    "gunicorn>=23.0.0",
    "llama-index>=0.14.2",
    "llama-index-embeddings-azure-openai>=0.4.1",
    "llama-index-llms-azure-openai>=0.4.1",
    "llama-index-readers-file>=0.5.4",
    "llama-index-readers-web>=0.5.3",
    "openai==1.108.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
    "uvicorn>=0.36.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D1",    # undocumented
    "D203",  # one blank line before class
    "D213",  # multi-line summary second line
    "TD001", # invalid todo tag
    "TD002", # missing todo author
    "TD003", # missing todo link
    "PD011", # pandas use of dot values
]
unfixable = [
    "F401", # unused import
    "F841", # unused variable
]

logger-objects = ["src.library.logger.LOGGER"]

[tool.ruff.lint.pylint]
max-args = 6
